renamed some of the files and added some status to both listing creation and application
