+++ src/components/ListingDetails/WorkerView.tsx
@@ -3,7 +3,7 @@ import UserDeployments from "@/components/ListingDetails/UserDeployments";
 import DeploymentTabs from "@/components/ListingDetails/DeploymentTabs";
 import { ImArrowUpRight2 } from "react-icons/im";
 import TransactionCard from "@/components/ListingDetails/TransactionCard";
-import IListing from "@/interfaces/listingResponse";
+import IListing from "@/interfaces/ListingResponse";
 import { short } from "@/utils/Common";
 
 type Props = {
