+++ src/enums/JobStatus.ts
@@ -1,0 +1,10 @@
+enum JobStatus {
+  PENDING,
+  APPLICATION,
+  ACTIVE,
+  CANCELED,
+  COMPLETED,
+  DISPUTED,
+}
+
+export default JobStatus;
