+++ src/interfaces/ApplicantResponse.ts
