+++ src/interfaces/CategoryResponse.ts
