+++ src/interfaces/FileInfo.ts
