+++ src/interfaces/ListingResponse.ts
