+++ src/utils/Categories.ts
@@ -1,5 +1,5 @@
 import axios from "axios";
-import ICategory from "@/interfaces/categoryResponse";
+import ICategory from "@/interfaces/CategoryResponse";
 
 const fetchCategories = async () => {
   let categoryData: ICategory[] = [];
